Add files via upload
You are given a list of orders placed by various customers along with the order placed date. Each order contains one or more items. Write a code to find the two most frequently purchased items by the customers in the last ‘30’ days.

 

Example:

    Order ‘1’ was placed for the item “Watch” 2 months ago.
    Order ‘2’ was placed for the items “Mobile”, “Watch” and “Shoe” 20 days ago.
    Order ‘3’ was placed for the item “Mobile” 10 days ago.
    Order ‘4’ was placed for the item “Shoe” 5 days ago.

 

Order 1 was placed 2 months ago whereas Orders 2, 3, and 4 were placed in the last month. In this case, “Mobile” and “Shoe” should be returned as output are the two most frequently purchased items in the last ‘30’ days.